requirements.txt
anyio==3.6.2
asgiref==3.5.2
click==8.1.3
fastapi==0.88.0
greenlet==2.0.1
h11==0.14.0
idna==3.4
loguru==0.6.0
psycopg2-binary==2.9.2
pydantic==1.8.2
PyJWT==1.7.1
sniffio==1.3.0
SQLAlchemy==1.4.27
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.17.6